extract password strength logic to standalone component
